Updates per Jason's feedback
